requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
insightface  # 移除版本限制，使用最新版
faiss-cpu==1.11.0
opencv-python==4.8.1.78
pillow==10.4.0
numpy>=1.26.0
scikit-learn==1.3.2
werkzeug==3.0.1
python-multipart==0.0.6
sqlalchemy==2.0.23
onnxruntime  # insightface 的依赖